kruskal/requirements.txt
streamlit>=1.32.0
networkx>=3.1
folium>=0.15.1
streamlit-folium>=0.14.0
pymongo>=4.6.1
python-dotenv>=1.0.1
matplotlib>=3.7.1
pandas>=2.1.0
plotly>=5.18.0
